+++ changelog.md
@@ -1,4 +1,4 @@
 ### General
 
 - Added carpet rule `strictEndPortal`
-- Support for 26.1-snapshot-3
+- New icon