+++ changelog.md
@@ -1,4 +1,3 @@
 ### General
 
-- Added carpet rule `strictEndPortal`
-- New icon
+- Improved command feedback for `/endermanBlockList` and fixed some i18n problems